% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfold.R
\name{nfold_breaks}
\alias{nfold_breaks}
\title{N-fold Breaks}
\usage{
nfold_breaks(n = 5)
}
\arguments{
\item{n}{A numeric scalar of the desired number of breaks.}
}
\value{
A function to calculate the breaks from a vector.
}
\description{
N-fold Breaks
}
\examples{
nfold_breaks()(c(-3 / 4, -2 / 3, -1 / 2, 0, 1, 2, 3))
}
