% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfold.R
\name{prop2nfold}
\alias{prop2nfold}
\alias{nfold2prop}
\title{Proportional Change to N-fold Change}
\usage{
prop2nfold(x)

nfold2prop(x)
}
\arguments{
\item{x}{A double vector of the proportional (n-fold) change.}
}
\value{
A double vector of the n-fold (proportional) change.
}
\description{
Proportional Change to N-fold Change
}
\section{Functions}{
\itemize{
\item \code{nfold2prop}: N-fold Change to Proportional Change
}}

\examples{
prop2nfold(c(-1, -0.5, 0, 1, 2))
nfold2prop(prop2nfold(c(-1, -0.5, 0, 1, 2)))
}
