% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfold.R
\name{scale_x_nfold}
\alias{scale_x_nfold}
\alias{scale_y_nfold}
\title{N-fold Position Scales}
\usage{
scale_x_nfold(...)

scale_y_nfold(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{scale_(x|y)_continuous()}}
}
\description{
scale_x_nfold() and scale_y_nfold() are variants of \code{\link{scale_x_continuous}()}
that set the trans argument to transform proportional change to n-fold change.
}
