% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{set_colors}
\alias{set_colors}
\title{Sets colors}
\usage{
set_colors(values = get_colors())
}
\arguments{
\item{values}{character vector of colors}
}
\value{
invisible character vector of colors
}
\description{
Sets color palette for poiscon
}

