% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-Golder.R
\name{theme_Golder}
\alias{theme_Golder}
\title{Theme Golder}
\usage{
theme_Golder(base_size = 10, base_family = "Helvetica")
}
\arguments{
\item{base_size}{an integer scalar indicating the base font size.}

\item{base_family}{a character scalar indicating the base font family.}
}
\value{
A ggplot2 theme.
}
\description{
A ggplot2 theme for Golder plots
}
