% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-Poisson.R
\name{theme_Poisson}
\alias{theme_Poisson}
\title{Theme Poisson}
\usage{
theme_Poisson(base_size = 10, base_family = "")
}
\arguments{
\item{base_size}{an integer scalar indicating the base font size.}

\item{base_family}{a character scalar indicating the base font family.}
}
\value{
A gg theme object.
}
\description{
A gg theme for Poisson style plots.
}
\details{
To use, ensure ggplot2 is installed and then call
\code{theme_set(theme_Poisson())}.
}

